Made a change to FastSet.Add(ICollection<T>)
No longer checking for the existence of members before adding a collection to a set.
It is easier for a client to just "add what it has" than to figure out the difference between what the client has and what's already in the Population.
The client can check its list against the modified set's membership to see what is missing.